+++ database/seeders/DatabaseSeeder.php
@@ -12,6 +12,6 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        $this->call(binance::class);
+        $this->call(users::class);
     }
 }
