--- storage/app/public/ALGO-USD.csv
@@ -1,102 +1,0 @@
-Date,Open,High,Low,Close,Adj Close,Volume
-2022-12-30,0.165134,0.169008,0.160899,0.169008,0.169008,41936642
-2022-12-31,0.169055,0.174602,0.167819,0.172818,0.172818,46000196
-2023-01-01,0.172810,0.178576,0.171996,0.177466,0.177466,36884680
-2023-01-02,0.177473,0.182622,0.174552,0.181194,0.181194,39522483
-2023-01-03,0.181203,0.185339,0.180468,0.184302,0.184302,42799437
-2023-01-04,0.184303,0.191343,0.183980,0.187093,0.187093,52324895
-2023-01-05,0.187080,0.190731,0.182760,0.183544,0.183544,43347762
-2023-01-06,0.183571,0.189930,0.181428,0.189497,0.189497,44074050
-2023-01-07,0.189496,0.195472,0.188530,0.194386,0.194386,42834111
-2023-01-08,0.194324,0.199651,0.192284,0.199550,0.199550,38236365
-2023-01-09,0.199472,0.207882,0.199396,0.200688,0.200688,85397892
-2023-01-10,0.200599,0.204710,0.198820,0.204071,0.204071,50399010
-2023-01-11,0.204059,0.210734,0.197617,0.210734,0.210734,53987029
-2023-01-12,0.210797,0.212838,0.205301,0.212526,0.212526,83284144
-2023-01-13,0.212540,0.222206,0.209107,0.221019,0.221019,69730976
-2023-01-14,0.221009,0.239102,0.220998,0.229522,0.229522,146305142
-2023-01-15,0.229580,0.240335,0.225049,0.240335,0.240335,107565151
-2023-01-16,0.240355,0.241336,0.226489,0.234289,0.234289,98057323
-2023-01-17,0.234326,0.235364,0.229497,0.229497,0.229497,66725576
-2023-01-18,0.229521,0.232480,0.209189,0.209594,0.209594,102596761
-2023-01-19,0.209564,0.217985,0.209480,0.217528,0.217528,57599865
-2023-01-20,0.217510,0.239863,0.216263,0.239776,0.239776,81520520
-2023-01-21,0.239762,0.245262,0.234036,0.239667,0.239667,100722878
-2023-01-22,0.239618,0.254386,0.236634,0.246711,0.246711,121941539
-2023-01-23,0.246712,0.254221,0.246521,0.250303,0.250303,99493916
-2023-01-24,0.250249,0.259150,0.235680,0.237040,0.237040,110474128
-2023-01-25,0.237049,0.249084,0.230904,0.245368,0.245368,87129537
-2023-01-26,0.245197,0.248541,0.242237,0.245808,0.245808,70905794
-2023-01-27,0.245803,0.258505,0.240456,0.254728,0.254728,87973627
-2023-01-28,0.254744,0.263761,0.251982,0.257302,0.257302,84403661
-2023-01-29,0.257436,0.264464,0.255396,0.262193,0.262193,71822187
-2023-01-30,0.262212,0.262374,0.237563,0.239771,0.239771,85126947
-2023-01-31,0.239777,0.244329,0.236731,0.242206,0.242206,61930182
-2023-02-01,0.242161,0.251926,0.232595,0.251032,0.251032,96058386
-2023-02-02,0.251049,0.267764,0.251049,0.255414,0.255414,106102449
-2023-02-03,0.255375,0.272973,0.252195,0.270928,0.270928,132688284
-2023-02-04,0.270897,0.273587,0.266294,0.267021,0.267021,78614505
-2023-02-05,0.267054,0.270078,0.252879,0.257102,0.257102,83152126
-2023-02-06,0.257133,0.263153,0.255147,0.255421,0.255421,66235465
-2023-02-07,0.255363,0.285231,0.254984,0.279158,0.279158,194585706
-2023-02-08,0.279191,0.296965,0.278117,0.290014,0.290014,213731969
-2023-02-09,0.290033,0.290309,0.249279,0.255304,0.255304,197237440
-2023-02-10,0.255305,0.263086,0.250254,0.252010,0.252010,115126614
-2023-02-11,0.252008,0.257216,0.249367,0.257189,0.257189,62079971
-2023-02-12,0.257185,0.259801,0.245925,0.248337,0.248337,68975253
-2023-02-13,0.248338,0.249003,0.235428,0.244658,0.244658,90996780
-2023-02-14,0.244624,0.254939,0.240782,0.253670,0.253670,86028993
-2023-02-15,0.253630,0.273444,0.251761,0.272708,0.272708,106214843
-2023-02-16,0.272644,0.280072,0.258004,0.258604,0.258604,145083641
-2023-02-17,0.258570,0.275149,0.258395,0.271575,0.271575,99575419
-2023-02-18,0.271542,0.277587,0.269517,0.273514,0.273514,70508283
-2023-02-19,0.273471,0.282468,0.269359,0.278160,0.278160,106011320
-2023-02-20,0.278298,0.290535,0.270867,0.284993,0.284993,136180970
-2023-02-21,0.285123,0.292958,0.274580,0.278345,0.278345,121387968
-2023-02-22,0.278347,0.279338,0.262617,0.271323,0.271323,99013699
-2023-02-23,0.271320,0.278776,0.266718,0.268714,0.268714,86818704
-2023-02-24,0.268766,0.269409,0.252783,0.256846,0.256846,79896942
-2023-02-25,0.256837,0.257660,0.245758,0.251300,0.251300,52253817
-2023-02-26,0.251296,0.256162,0.249552,0.255726,0.255726,37908316
-2023-02-27,0.255720,0.261256,0.250900,0.255232,0.255232,51910791
-2023-02-28,0.255168,0.255602,0.244090,0.244947,0.244947,50579740
-2023-03-01,0.244960,0.254219,0.244208,0.251219,0.251219,50986141
-2023-03-02,0.251227,0.251904,0.240808,0.242597,0.242597,48090504
-2023-03-03,0.242552,0.242616,0.225207,0.233446,0.233446,69471117
-2023-03-04,0.233447,0.234187,0.222578,0.225494,0.225494,33754152
-2023-03-05,0.225495,0.231275,0.224303,0.227186,0.227186,38013563
-2023-03-06,0.227181,0.230752,0.220173,0.225965,0.225965,62268405
-2023-03-07,0.225954,0.228006,0.213966,0.217646,0.217646,62127121
-2023-03-08,0.217564,0.217918,0.200867,0.202328,0.202328,61164379
-2023-03-09,0.202252,0.206922,0.189176,0.193574,0.193574,72973286
-2023-03-10,0.193579,0.195157,0.186182,0.189342,0.189342,101190349
-2023-03-11,0.189327,0.194786,0.181569,0.189221,0.189221,83489205
-2023-03-12,0.189201,0.203103,0.184403,0.202391,0.202391,71141684
-2023-03-13,0.202412,0.215821,0.195122,0.212931,0.212931,110748843
-2023-03-14,0.212925,0.227042,0.207357,0.222731,0.222731,119645575
-2023-03-15,0.222833,0.226253,0.200296,0.204872,0.204872,93774828
-2023-03-16,0.204979,0.210243,0.202703,0.209169,0.209169,51004142
-2023-03-17,0.209182,0.227241,0.207596,0.227241,0.227241,69940667
-2023-03-18,0.227232,0.229938,0.215365,0.215678,0.215678,82270493
-2023-03-19,0.215719,0.225130,0.215719,0.221018,0.221018,63720019
-2023-03-20,0.221057,0.225236,0.209613,0.210182,0.210182,79112674
-2023-03-21,0.210263,0.225653,0.206721,0.224551,0.224551,76462590
-2023-03-22,0.224461,0.240103,0.210754,0.214062,0.214062,211957606
-2023-03-23,0.214034,0.222581,0.212198,0.219691,0.219691,76021343
-2023-03-24,0.219656,0.220090,0.207058,0.209219,0.209219,63594104
-2023-03-25,0.209212,0.211298,0.202995,0.206541,0.206541,40651505
-2023-03-26,0.206543,0.211630,0.205643,0.210829,0.210829,30772734
-2023-03-27,0.210819,0.210884,0.196433,0.200274,0.200274,53280488
-2023-03-28,0.200251,0.204066,0.197409,0.202006,0.202006,50531386
-2023-03-29,0.201990,0.237527,0.201858,0.234327,0.234327,279749230
-2023-03-30,0.234348,0.238503,0.221698,0.224584,0.224584,131918255
-2023-03-31,0.224588,0.227625,0.218469,0.225408,0.225408,74228214
-2023-04-01,0.225482,0.227853,0.219442,0.222180,0.222180,46137695
-2023-04-02,0.223216,0.224335,0.211790,0.214095,0.214095,44963360
-2023-04-03,0.214097,0.219000,0.208308,0.214364,0.214364,63103831
-2023-04-04,0.214341,0.220408,0.212243,0.218122,0.218122,48671618
-2023-04-05,0.218123,0.223790,0.216156,0.222892,0.222892,62049037
-2023-04-06,0.222905,0.222905,0.216192,0.217775,0.217775,51920289
-2023-04-07,0.217811,0.219615,0.213635,0.219317,0.219317,41784835
-2023-04-08,0.219304,0.220808,0.214881,0.216299,0.216299,41570807
-2023-04-09,0.216265,0.218121,0.213516,0.216851,0.216851,34754910
